rename: spec을 별도의 패키지로 이동
