+++ src/main/java/com/example/masterplanbbe/domain/exam/entity/ExamDetail.java
@@ -1,6 +1,7 @@
 package com.example.masterplanbbe.domain.exam.entity;
 
 import com.example.masterplanbbe.common.domain.FullAuditEntity;
+import com.example.masterplanbbe.domain.spec.entity.Spec;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
