+++ src/main/java/com/example/masterplanbbe/domain/exam/entity/Subject.java
@@ -1,7 +1,7 @@
 package com.example.masterplanbbe.domain.exam.entity;
 
-import com.example.masterplanbbe.common.annotation.NonNull;
 import com.example.masterplanbbe.common.domain.FullAuditEntity;
+import com.example.masterplanbbe.domain.spec.entity.Spec;
 import jakarta.persistence.*;
 import lombok.*;
