+++ src/main/java/com/example/masterplanbbe/domain/spec/repository/SpecRepository.java
@@ -1,6 +1,6 @@
-package com.example.masterplanbbe.domain.exam.repository;
+package com.example.masterplanbbe.domain.spec.repository;
 
-import com.example.masterplanbbe.domain.exam.entity.Spec;
+import com.example.masterplanbbe.domain.spec.entity.Spec;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SpecRepository extends JpaRepository<Spec, Long> {
