+++ src/main/java/com/example/masterplanbbe/domain/spec/request/SpecCreateRequest.java
@@ -1,7 +1,7 @@
-package com.example.masterplanbbe.domain.exam.request;
+package com.example.masterplanbbe.domain.spec.request;
 
 import com.example.masterplanbbe.domain.exam.entity.ExamDetail;
-import com.example.masterplanbbe.domain.exam.entity.Spec;
+import com.example.masterplanbbe.domain.spec.entity.Spec;
 import com.example.masterplanbbe.domain.exam.enums.Category;
 import com.example.masterplanbbe.domain.exam.enums.CertificationType;
